+++ latch/ldata/_transfer/utils.py
@@ -2,12 +2,24 @@ import os
 import time
 from typing import Any, Dict, Optional
 
-from gql.gql import DocumentNode
+import requests
+import requests.adapters
 from gql.transport.exceptions import TransportClosed, TransportServerError
+from graphql.language import DocumentNode
 from latch_sdk_gql import JsonValue
 from latch_sdk_gql.execute import execute
 
-from latch_cli import tinyrequests
+http_session = requests.Session()
+
+_adapter = requests.adapters.HTTPAdapter(
+    max_retries=requests.adapters.Retry(
+        status_forcelist=[429, 500, 502, 503, 504],
+        backoff_factor=1,
+        allowed_methods=["GET", "PUT", "POST"],
+    )
+)
+http_session.mount("https://", _adapter)
+http_session.mount("http://", _adapter)
 
 
 # todo(rahul): move this function into latch_sdk_gql.execute
@@ -21,6 +33,7 @@ def query_with_retry(
     Send GraphQL query request. Retry on Server or Connection failures.
     Implements exponential backoff between retries
     """
+    err = None
     attempt = 0
     while attempt < num_retries:
         attempt += 1
@@ -34,7 +47,10 @@ def query_with_retry(
             # todo(rahul): tune the sleep interval based on the startup time of the vacuole
             time.sleep(2**attempt * 5)
 
-    raise err
+    if err is not None:
+        raise err
+
+    raise RuntimeError("gql retries exceeded")
 
 
 def get_max_workers() -> int: