+++ CHANGELOG.md
@@ -16,12 +16,20 @@ Types of changes
 
 # Latch SDK Changelog
 
-## v2.53.2 - 2024-10-02
+## v2.53.3 - 2024-10-02
 
 ### Added
 
 * Nextflow
   - Add support for uploading command logs after task execution
+
+## v2.53.2 - 2024-10-02
+
+### Changed
+
+* All requests done by `latch cp` now
+  - use connection pooling / reuses
+  - have automatic retries with backoff
 
 ## v2.53.1 - 2024-10-02